+++ map-matcher-library/src/test/java/nu/ndw/nls/routingmapmatcher/isochrone/v2/exploration/ExploreTimeLimitTest.java
@@ -26,6 +26,14 @@ class ExploreTimeLimitTest {
     }
 
     @Test
+    void constructor() {
+        exploreTimeLimit = new ExploreTimeLimit<>(LIMIT, true);
+
+        assertThat(exploreTimeLimit.isApplyLimitToParent()).isTrue();
+        assertThat(exploreTimeLimit.getLimit()).isEqualTo(LIMIT);
+    }
+
+    @Test
     void isInLimit() {
         IsochroneLabel label = new IsochroneLabel(0, 5, 5, null, (long) (LIMIT - 0.1), 0.0, 0.0);
 
@@ -37,5 +45,12 @@ class ExploreTimeLimitTest {
         IsochroneLabel label = new IsochroneLabel(0, 5, 5, null, (long) LIMIT, 0.0, 0.0);
 
         assertThat(exploreTimeLimit.isInLimit(label, encodingManager)).isFalse();
+    }
+
+    @Test
+    void debug() {
+        IsochroneLabel label = new IsochroneLabel(0, 5, 5, null, (long) (LIMIT - 0.1), 0.0, 0.0);
+
+        assertThat(exploreTimeLimit.debug(label, encodingManager)).isEqualTo("ExploreTimeLimit{limit=3600.0, time=3599.0, reached=false}");
     }
 }