+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 Here we write upgrade notes. It's a team effort to make them as straightforward as possible.
+## [21.2.0] - 2026-06-10
+- Added ExploreLimitComposite
+- Exploration limits now support checking against the isochroneLabel parent. This is relevant if you want to include the road section where the limit is reached in the isochrone calculation.
+- Improved debugging capabilities.
 
 ## [21.1.0] - 2026-06-10
  - Added [AbstractDijkstraIsochroneAlgorithm.java](map-matcher-library/src/main/java/nu/ndw/nls/routingmapmatcher/isochrone/v2/algorithm/AbstractDijkstraIsochroneAlgorithm.java) to replace [AbstractShortestPathTree.java](map-matcher-library/src/main/java/nu/ndw/nls/routingmapmatcher/isochrone/algorithm/AbstractShortestPathTree.java) and [IsochroneByTimeDistanceAndWeight.java](map-matcher-library/src/main/java/nu/ndw/nls/routingmapmatcher/isochrone/algorithm/IsochroneByTimeDistanceAndWeight.java)